removed cppxml and replaced code using it with RapidXml

+++ source/winlame/old_ui/PresetsPage.cpp
@@ -84,8 +84,8 @@ void PresetsPage::OnEnterPage()
    CString cszText(MAKEINTRESOURCE(IDS_PRESETS_CUSTOM_SETTINGS));
    listbox.AddString(cszText);
 
-   int max = m_presetManager.getPresetCount();
-   for(int i=0; i<max; i++)
+   size_t max = m_presetManager.getPresetCount();
+   for(size_t i=0; i<max; i++)
       listbox.AddString(m_presetManager.getPresetName(i).c_str());
 
    listbox.SetCurSel(lastindex);
